probando git fetch

+++ casting.py
@@ -1,6 +1,6 @@
 print("Ingresa un numero")
 num1 = int(input())
 print("Ingresa un numero")
-num2 = int(input())
+num2 = float(input())
 res = num1 + num2
 print("El resultado es: " + str (res))
